Part 1

+++ src/y2019/day2/Part2.java
@@ -4,16 +4,16 @@ import utils.Problem;
 import y2019.intcode.Machine;
 import y2019.intcode.Parser;
 
-public class Part2 extends Problem<Integer> {
+public class Part2 extends Problem<Long> {
     private Parser parser;
 
-    public Part2(String fileName, int exp) {
+    public Part2(String fileName, long exp) {
         super(exp);
 
         parser = new Parser(fileName);
     }
 
-    private int runMachine(int[] prog, int noun, int verb) {
+    private long runMachine(long[] prog, int noun, int verb) {
         var mach = new Machine(prog);
 
         mach.set(1, noun);
@@ -26,10 +26,10 @@ public class Part2 extends Problem<Integer> {
         return mach.get(0);
     }
 
-    public Integer run() {
+    public Long run() {
         var prog = parser.parse();
         var target = 19690720;
-        var answer = 0;
+        var answer = 0L;
 
         for (var noun = 0; answer == 0 && noun < 100; noun += 1) {
             for (var verb = 0; answer == 0 && verb < 100; verb += 1) {
