+++ src/y2019/intcode/Parser.java
@@ -1,19 +1,19 @@
 package y2019.intcode;
 
-public class Parser extends utils.Parser<int[]> {
+public class Parser extends utils.Parser<long[]> {
     public Parser(String fileName) {
         super(fileName);
     }
 
-    public int[] parse() {
+    public long[] parse() {
         try {
             var lines = readLines();
             var line = lines.get(0);
             var parts = line.split(",");
-            var prog = new int[parts.length];
+            var prog = new long[parts.length];
 
             for (var ndx = 0; ndx < parts.length; ndx += 1) {
-                prog[ndx] = Integer.parseInt(parts[ndx]);
+                prog[ndx] = Long.parseLong(parts[ndx]);
             }
 
             return prog;
